[Housekeeping] Cleanup upload view code.
